4bitQ/pyproject.toml
[tool.poetry]
name = "4bitq"
version = "0.1.0"
description = ""
authors = ["BREW <IRACK000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
llama-cpp-python = "^0.2.75"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
